refactor: split up `main` and `renderer` files
